[domain] Move Action from lib
