+++ lib/src/main/scala/net/kemitix/thorp/lib/PlanBuilder.scala
@@ -2,7 +2,7 @@ package net.kemitix.thorp.lib
 
 import net.kemitix.thorp.config.Config
 import net.kemitix.thorp.console._
-import net.kemitix.thorp.lib.Action._
+import net.kemitix.thorp.domain.Action._
 import net.kemitix.thorp.domain._
 import net.kemitix.thorp.filesystem.{FileSystem, Hasher}
 import net.kemitix.thorp.storage.Storage
