+++ lib/src/main/scala/net/kemitix/thorp/lib/SequencedAction.scala
@@ -1,5 +1,7 @@
 package net.kemitix.thorp.lib
 
+import net.kemitix.thorp.domain.Action
+
 final case class SequencedAction(
     action: Action,
     index: Int
