+++ lib/src/test/scala/net/kemitix/thorp/lib/ActionGeneratorSuite.scala
@@ -1,7 +1,7 @@
 package net.kemitix.thorp.lib
 
 import net.kemitix.thorp.config._
-import net.kemitix.thorp.lib.Action.{DoNothing, ToCopy, ToUpload}
+import net.kemitix.thorp.domain.Action.{DoNothing, ToCopy, ToUpload}
 import net.kemitix.thorp.domain.HashType.MD5
 import net.kemitix.thorp.domain._
 import net.kemitix.thorp.filesystem.{FileSystem, Resource}
