+++ lib/src/test/scala/net/kemitix/thorp/lib/PlanBuilderTest.scala
@@ -10,7 +10,7 @@ import net.kemitix.thorp.config.{
   ConfigurationBuilder
 }
 import net.kemitix.thorp.console._
-import net.kemitix.thorp.lib.Action.{DoNothing, ToCopy, ToDelete, ToUpload}
+import net.kemitix.thorp.domain.Action.{DoNothing, ToCopy, ToDelete, ToUpload}
 import net.kemitix.thorp.domain.HashType.MD5
 import net.kemitix.thorp.domain._
 import net.kemitix.thorp.filesystem.{Hasher, _}
